refactor(ai): isolate Responses invoke logic
