+++ cli/azd/extensions/azure.ai.agents/internal/cmd/invoke.go
@@ -979,156 +979,6 @@ func doLocalRequestWithRetry(
 	}
 }
 
-func (a *InvokeAction) responsesLocal(ctx context.Context) error {
-	port := a.flags.port
-
-	body, bodyLabel, err := a.resolveBody()
-	if err != nil {
-		return err
-	}
-
-	msg := string(body)
-
-	// Open azd client for session/conversation persistence.
-	var azdClient *azdext.AzdClient
-	if c, err := azdext.NewAzdClient(); err == nil {
-		azdClient = c
-		defer azdClient.Close()
-	}
-
-	agentKey := resolveLocalAgentKey(ctx, azdClient, a.serviceNameSelector(), a.noPrompt)
-
-	// Resolve local session and conversation IDs (always generated locally).
-	var sid, convID string
-	if azdClient != nil {
-		sid, err = resolveStoredID(
-			ctx, azdClient, agentKey, a.flags.session, a.flags.newSession, "sessions", true,
-		)
-		if err != nil {
-			log.Printf("invoke local: failed to resolve session ID: %v", err)
-		}
-		convID, err = resolveStoredID(
-			ctx, azdClient, agentKey, a.flags.conversation, a.flags.newConversation, "conversations", true,
-		)
-		if err != nil {
-			log.Printf("invoke local: failed to resolve conversation ID: %v", err)
-		}
-	}
-
-	raw := a.flags.outputFmt == outputRaw
-	if !raw {
-		fmt.Printf("Target:       localhost:%d (local)\n", port)
-		fmt.Printf("Message:      %s\n", bodyLabel)
-		printSessionStatus("Session:      ", sid)
-		fmt.Printf("Conversation: %s\n\n", convID)
-	}
-
-	reqBody := map[string]any{
-		"input": msg,
-	}
-	if sid != "" {
-		reqBody["session_id"] = sid
-	}
-	if convID != "" {
-		reqBody["conversation"] = map[string]string{"id": convID}
-	}
-
-	payload, err := json.Marshal(reqBody)
-	if err != nil {
-		return fmt.Errorf("failed to marshal request: %w", err)
-	}
-
-	reqURL := fmt.Sprintf("http://localhost:%d/responses", port)
-	newReq := func() (*http.Request, error) {
-		req, err := http.NewRequestWithContext(ctx, http.MethodPost, reqURL, bytes.NewReader(payload))
-		if err != nil {
-			return nil, err
-		}
-		applyCustomHeaders(req, a.clientHeaders)
-		req.Header.Set("Content-Type", "application/json")
-		applyLocalUserIdentityHeader(req, &a.flags.userIdentityFlags)
-		applyLocalCallIDHeader(req, a.flags.callID)
-		if raw {
-			// Disable Go's transparent gzip handling so the dumped headers and
-			// body match what the server actually sent on the wire.
-			req.Header.Set("Accept-Encoding", "identity")
-		}
-		return req, nil
-	}
-
-	client := &http.Client{Timeout: a.httpTimeout()}
-	invokeStart := time.Now()
-	resp, err := doLocalRequestWithRetry(ctx, client, newReq)
-	if err != nil {
-		return fmt.Errorf(
-			"could not connect to localhost:%d -- is the agent running? Start it with: azd ai agent run",
-			port,
-		)
-	}
-	ttfb := time.Since(invokeStart)
-	defer resp.Body.Close()
-
-	if raw {
-		// Stream the body verbatim to stdout (avoids buffering large responses).
-		if dumpErr := writeRawResponse(os.Stdout, resp); dumpErr != nil {
-			return dumpErr
-		}
-		if resp.StatusCode >= 400 {
-			return fmt.Errorf(
-				"POST %s failed with HTTP %d: %s",
-				reqURL, resp.StatusCode, resp.Status,
-			)
-		}
-		return nil
-	}
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return fmt.Errorf("failed to read response: %w", err)
-	}
-	totalDuration := time.Since(invokeStart)
-
-	if resp.StatusCode >= 400 {
-		if traceID := responseTraceID(resp); traceID != "" {
-			fmt.Printf("Trace ID:     %s\n", traceID)
-		}
-		a.emitInvokeFailureNextStep(nextstep.InvokeLocal, "", "")
-		return fmt.Errorf(
-			"POST %s failed with HTTP %d: %s\n%s",
-			reqURL, resp.StatusCode, resp.Status, string(respBody),
-		)
-	}
-
-	var result map[string]any
-	if err := json.Unmarshal(respBody, &result); err != nil {
-		// Not JSON -- just print raw response
-		fmt.Println(string(respBody))
-		printInvokeTiming(os.Stdout, totalDuration, ttfb)
-		a.emitInvokeSuccessNextStep(nextstep.InvokeLocal, "")
-		return nil
-	}
-
-	if err := printAgentResponse(result, "local"); err != nil {
-		return err
-	}
-	printInvokeTiming(os.Stdout, totalDuration, ttfb)
-	a.emitInvokeSuccessNextStep(nextstep.InvokeLocal, "")
-	return nil
-}
-
-// remoteContext holds the resolved inputs for a remote (Foundry) invoke.
-// In ephemeral mode (--agent-endpoint) the project endpoint / agent name /
-// api-version come from the parsed URL.
-//
-// agentKey is the persistence key used by the global UserConfig store. It is
-// non-empty whenever session/conversation IDs should be saved or resumed:
-//   - project mode: derived from AGENT_{SVC}_ENDPOINT
-//   - ephemeral mode: derived from the parsed --agent-endpoint URL
-//     (independent of api-version / trailing slash / fragment)
-//
-// In standalone mode (no parent azd daemon, e.g. running the extension binary
-// directly outside an azd command) azdClient is nil and persistence helpers
-// no-op. agentKey may still be non-empty in that case.
 type remoteContext struct {
 	name                               string
 	serviceName                        string
@@ -2067,55 +1917,6 @@ func handleInvocationLRO(
 	}
 }
 
-// createConversation creates a new Foundry conversation for multi-turn memory.
-func createConversation(
-	ctx context.Context,
-	projectEndpoint, agentName, bearerToken, apiVersion string,
-	options *agent_api.SessionRequestOptions,
-) (string, error) {
-	if apiVersion == "" {
-		apiVersion = DefaultAgentAPIVersion
-	}
-	convURL := fmt.Sprintf(
-		"%s/agents/%s/endpoint/protocols/openai/conversations?api-version=%s",
-		projectEndpoint, agentName, url.QueryEscape(apiVersion),
-	)
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, convURL, bytes.NewReader([]byte("{}")))
-	if err != nil {
-		return "", err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+bearerToken)
-	options.ApplyHeaders(req.Header)
-
-	client := &http.Client{Timeout: 30 * time.Second}
-	resp, err := client.Do(req) //nolint:gosec // G704: endpoint is resolved from azd environment configuration
-	if err != nil {
-		return "", fmt.Errorf("POST %s failed: %w", convURL, err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode >= 400 {
-		respBody, _ := io.ReadAll(resp.Body)
-		return "", fmt.Errorf("POST %s failed with HTTP %d: %s\n%s", convURL, resp.StatusCode, resp.Status, string(respBody))
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-
-	var result map[string]any
-	if err := json.Unmarshal(body, &result); err != nil {
-		return "", err
-	}
-
-	if id, ok := result["id"].(string); ok {
-		return id, nil
-	}
-	return "", fmt.Errorf("conversation response missing 'id' field")
-}
-
 // responseTraceID returns the trace ID from the response, preferring x-request-id
 // and falling back to apim-request-id. If a header value is comma-folded (which
 // can happen when an intermediary like APIM combines duplicate headers per
@@ -2131,61 +1932,4 @@ func responseTraceID(resp *http.Response) string {
 		}
 	}
 	return ""
-}
-
-// printAgentResponse pretty-prints the output_text items from an agent response.
-func printAgentResponse(result map[string]any, title string) error {
-	// Check for agent-level errors (e.g., agent runtime failures)
-	if status, _ := result["status"].(string); status == "failed" {
-		if errObj, ok := result["error"].(map[string]any); ok {
-			msg, _ := errObj["message"].(string)
-			code, _ := errObj["code"].(string)
-			return fmt.Errorf("agent failed (%s): %s", code, msg)
-		}
-		return fmt.Errorf("agent returned failed status")
-	}
-
-	// Check for server-level errors (e.g., local agentserver: {"code": "server_error", "message": "..."})
-	if code, ok := result["code"].(string); ok && code != "" {
-		msg, _ := result["message"].(string)
-		return fmt.Errorf("agent error (%s): %s", code, msg)
-	}
-
-	outputItems, ok := result["output"].([]any)
-	if !ok {
-		// Try printing the whole response as formatted JSON
-		jsonBytes, _ := json.MarshalIndent(result, "", "  ")
-		fmt.Println(string(jsonBytes))
-		return nil
-	}
-
-	printed := false
-	for _, item := range outputItems {
-		itemMap, ok := item.(map[string]any)
-		if !ok {
-			continue
-		}
-		contentItems, ok := itemMap["content"].([]any)
-		if !ok {
-			continue
-		}
-		for _, content := range contentItems {
-			contentMap, ok := content.(map[string]any)
-			if !ok {
-				continue
-			}
-			if contentMap["type"] == "output_text" {
-				if text, ok := contentMap["text"].(string); ok {
-					fmt.Printf("[%s] %s\n", title, text)
-					printed = true
-				}
-			}
-		}
-	}
-
-	if !printed {
-		jsonBytes, _ := json.MarshalIndent(result, "", "  ")
-		fmt.Println(string(jsonBytes))
-	}
-	return nil
 }
